Improve epoch_read function
The types "uint64_t" and "unsigned long long" have the same width on
this platform. Making this change would not be necessary if I did not
have those "unsigned long long" casts for the return values.

+++ kernel/misc/epoch.c
@@ -82,7 +82,7 @@ static uint64_t slow_read(void)
 
 unsigned long long epoch_read(void)
 {
-	unsigned long long retval;
+	uint64_t retval;
 
 	if ((retval = fast_read(60000)) != 0)
 		return (unsigned long long)retval;
